+++ packages/atom.io/__tests__/experimental/realtime/realtime-continuity.test.tsx
@@ -1,8 +1,6 @@
 import { act, waitFor } from "@testing-library/react"
 import * as AtomIO from "atom.io"
-import { findRelationsInStore } from "atom.io/data"
-import type { Store } from "atom.io/internal"
-import { actUponStore, arbitrary, getFromStore } from "atom.io/internal"
+import { actUponStore, arbitrary } from "atom.io/internal"
 import * as AR from "atom.io/react"
 import * as RT from "atom.io/realtime"
 import * as RTR from "atom.io/realtime-react"
@@ -13,22 +11,6 @@ import { SetRTX } from "atom.io/transceivers/set-rtx"
 import * as React from "react"
 
 import { throwUntil } from "../../__util__/waiting"
-
-function prefixLogger(store: Store, prefix: string) {
-	store.loggers[0] = new AtomIO.AtomIOLogger(`info`, undefined, {
-		info: (...args) => {
-			console.info(prefix, ...args)
-		},
-		warn: (...args) => {
-			console.warn(prefix, ...args)
-		},
-		error: (...args) => {
-			console.error(prefix, ...args)
-		},
-	})
-}
-
-// AtomIO.getState(RTC.myIdState)
 
 describe(`synchronizing transactions`, () => {
 	const scenario = () => {
@@ -56,25 +38,10 @@ describe(`synchronizing transactions`, () => {
 		return RTTest.multiClient({
 			port: 5465,
 			server: ({ socket, silo: { store } }) => {
-				// const userKeyState = findInStore(
-				// 	RTS.usersOfSockets.states.userKeyOfSocket,
-				// 	socket.id,
-				// 	store,
-				// )
-				// const userKey = getFromStore(userKeyState, store)
-				// prefixLogger(store, `server`)
-				// socket.onAny((event, ...args) => {
-				// 	console.log(`🛰 `, userKey, event, ...args)
-				// })
-				// socket.onAnyOutgoing((event, ...args) => {
-				// 	console.log(`🛰  >>`, userKey, event, ...args)
-				// })
-
 				const syncContinuity = RTS.realtimeContinuitySynchronizer({
 					socket,
 					store,
 				})
-
 				syncContinuity(countContinuity)
 			},
 			clients: {
@@ -83,14 +50,7 @@ describe(`synchronizing transactions`, () => {
 					const count = AR.useO(countState)
 					const store = React.useContext(AR.StoreContext)
 					const increment = actUponStore(incrementTX, arbitrary(), store)
-					// prefixLogger(store, `jane`)
-					// const { socket } = React.useContext(RTR.RealtimeContext)
-					// socket?.onAny((event, ...args) => {
-					// 	console.log(`📡 JANE`, event, ...args)
-					// })
-					// socket?.onAnyOutgoing((event, ...args) => {
-					// 	console.log(`📡 JANE >>`, event, ...args)
-					// })
+
 					return (
 						<>
 							<button
@@ -170,12 +130,6 @@ describe(`synchronizing transactions`, () => {
 
 describe(`mutable atoms in continuity`, () => {
 	const scenario = () => {
-		// const topicsOfSubjects = join({
-		// 	key: `topicsOfSubjects`,
-		// 	between: [`topic`, `subject`],
-		// 	cardinality: `1:n`,
-		// })
-
 		const myListAtom = AtomIO.atom<SetRTX<string>, SetRTXJson<string>>({
 			key: `myList`,
 			default: () => new SetRTX<string>(),
@@ -186,8 +140,7 @@ describe(`mutable atoms in continuity`, () => {
 
 		const addItemTX = AtomIO.transaction<(item: string) => void>({
 			key: `addItem`,
-			do: ({ get, set }, item) => {
-				const myList = get(myListAtom)
+			do: ({ set }, item) => {
 				set(myListAtom, (list) => list.add(item))
 			},
 		})
@@ -200,20 +153,8 @@ describe(`mutable atoms in continuity`, () => {
 		return Object.assign(
 			RTTest.singleClient({
 				port: 5475,
-				server: ({ socket, silo: { store } }) => {
-					const userKeyState = findRelationsInStore(
-						RTS.usersOfSockets,
-						socket.id,
-						store,
-					).userKeyOfSocket
-					const userKey = getFromStore(userKeyState, store)
-					prefixLogger(store, `server`)
-					socket.onAny((event, ...args) => {
-						console.log(`🛰 `, userKey, event, ...args)
-					})
-					socket.onAnyOutgoing((event, ...args) => {
-						console.log(`🛰  >>`, userKey, event, ...args)
-					})
+				server: ({ socket, enableLogging, silo: { store } }) => {
+					enableLogging()
 					const exposeContinuity = RTS.realtimeContinuitySynchronizer({
 						socket,
 						store,
@@ -222,38 +163,18 @@ describe(`mutable atoms in continuity`, () => {
 				},
 				client: () => {
 					RTR.useSyncContinuity(applicationContinuity)
-					// RTR.usePullMutable(myListAtom)
-					const store = React.useContext(AR.StoreContext)
-					prefixLogger(store, `client`)
-					const { socket } = React.useContext(RTR.RealtimeContext)
-					socket?.onAny((event, ...args) => {
-						console.log(`📡 CLIENT`, event, ...args)
-					})
-					socket?.onAnyOutgoing((event, ...args) => {
-						console.log(`📡 CLIENT >>`, event, ...args)
-					})
 					const myList = AR.useJSON(myListAtom)
 
-					return (
-						<>
-							<button
-								type="button"
-								// onClick={() => {
-								// 	myList.add(`hello`)
-								// }}
-								data-testid={`add`}
-							/>
-							<span data-testid={`state`}>{myList.members.length}</span>
-						</>
-					)
+					return <span data-testid={`state`}>{myList.members.length}</span>
 				},
 			}),
 			{ myListAtom, addItemTX },
 		)
 	}
 	test(`mutable initialization`, async () => {
-		const { client, server, teardown, myListAtom, addItemTX } = scenario()
+		const { client, server, teardown, addItemTX } = scenario()
 		const clientApp = client.init()
+		clientApp.enableLogging()
 		await waitFor(() => {
 			throwUntil(clientApp.socket.connected)
 		})
@@ -267,3 +188,32 @@ describe(`mutable atoms in continuity`, () => {
 		teardown()
 	})
 })
+
+// describe(`join in perspective`, () => {
+// 	const scenario = () => {
+// 		const countState = AtomIO.atom<number>({ key: `count`, default: 0 })
+// 		const userActionCountServerState = AtomIO.atom<number>({
+// 			key: `server:userActionCount`,
+// 			default: 0,
+// 		})
+
+// 		const incrementTX = AtomIO.transaction({
+// 			key: `increment`,
+// 			do: ({ set, env }) => {
+// 				const { name } = env().store.config
+// 				if (name === `SERVER`) {
+// 					set(userActionCountServerState, (c) => c + 1)
+// 				}
+// 				set(countState, (c) => c + 1)
+// 			},
+// 		})
+// 		const countContinuity = RT.continuity({
+// 			key: `count`,
+// 			config: (group) => group.add(countState).add(incrementTX),
+// 		})
+
+// 		return RTTest.multiClient({
+// 			port: 5485,
+// 			server: ({ socket, silo: { store } }) => {}
+// 		})
+// })
